refactor: move unseen messages theme into widget file
